# [ENHANCEMENT] Apply RabbitMQ classic queue version 2