Claimed to be better than version 1 cf:
- https://www.rabbitmq.com/blog/2023/05/17/rabbitmq-3.12-performance-improvements#classic-queues-massively-improved-classic-queues-v2-cqv2
- https://www.rabbitmq.com/blog/2023/05/17/rabbitmq-3.12-performance-improvements#312-cqv1-vs-cqv2

### Before (all classic queues v1)
![Screenshot from 2024-05-13 16-46-30](https://github.com/apache/james-project/assets/55171818/27dd9522-ec48-48ae-be7c-e5e3eb7d64e3)


### After (all classic queues v2)
![Screenshot from 2024-05-13 16-37-40](https://github.com/apache/james-project/assets/55171818/34489e17-125a-4f43-9042-5f4f4c6b436c)

Better performance overall (p99 805ms vs 914ms).
SELECT p99 1874ms vs 2049ms
UNSELECT p99 1420ms vs 1755ms