@quantranhong1999 2024-05-13
> IE ensure this is not a breaking change.

Sadly this is a breaking change (queue purge is needed).
I would add a configuration to avoid this breaking change e.g. for the older James versions.